Support external training database for test build and predict

+++ StrainAMR_build_test.py
@@ -101,14 +101,11 @@ def merge_all_proteins(indir,odir,t):
     os.system('cat '+indir+'/* >'+odir+'/merged_proteins_'+t+'.fa')
     return odir+'/merged_proteins_'+t+'.fa'
     
-def run_cdhit(pval,work_dir):
-    # Run CD-Hit on all training proteins
-    #os.system('cd-hit -i '+ptrain+' -d 0 -o '+work_dir+'/merge_train_cdhit -c 0.9 -n 5 -M 0')
-    # Run CD-Hit on test proteins
-    os.system('cd-hit-2d -i '+work_dir+'/merged_proteins_train.fa -i2 '+pval+' -d 0 -o '+work_dir+'/merge_test_cdhit -c 0.9 -n 5 -M 0 -T 0')
-    cls1=work_dir+'/merge_train_cdhit.clstr'
+def run_cdhit(pval,work_dir,train_merged_proteins):
+    # Run CD-Hit on test proteins against the training representatives
+    os.system('cd-hit-2d -i '+train_merged_proteins+' -i2 '+pval+' -d 0 -o '+work_dir+'/merge_test_cdhit -c 0.9 -n 5 -M 0 -T 0')
     cls2=work_dir+'/merge_test_cdhit.clstr'
-    return cls1,cls2
+    return cls2
 
 def output_pc_token_file(d,pdir,label2,ofile):
     dr={} # Strain prefix -> Tokens string
@@ -161,8 +158,8 @@ def output_pc_token_file(d,pdir,label2,ofile):
 
     
 
-def generate_tokens_from_cdhit(work_dir,label2):
-    f=open(work_dir+'/merge_train_cdhit.clstr','r')
+def generate_tokens_from_cdhit(train_cluster_file,work_dir,label2):
+    f=open(train_cluster_file,'r')
     #o=open(work_dir+'/pc_matches.txt','w+')
     dcls={} # Cls_ID -> proteins
     arr=[]
@@ -211,7 +208,7 @@ def generate_tokens_from_cdhit(work_dir,label2):
     output_pc_token_file(d2,work_dir+'/proteins_test',label2,work_dir+'/strains_test_pc_token.txt')
     
 
-def run_ps(intest,label,label2,drug,work_dir):
+def run_ps(intest,label,label2,drug,work_dir,train_dir):
     #o=open('tem.pheno','w+') 
     #o.write('ID\tAddress\t'+drug+'\n')
     #dtrain={} # Pre -> Genome dir
@@ -259,12 +256,23 @@ def run_ps(intest,label,label2,drug,work_dir):
     #exit()
             
     #os.system(script_dir+'/PhenotypeSeeker/phenotypeseeker modeling tem.pheno')
+    build_dir(work_dir+'/PS_out')
+
+    train_ps_dir = os.path.join(train_dir, 'PS_out')
+    model_path = os.path.join(train_ps_dir, f'log_reg_model_{drug}.pkl')
+    if not os.path.exists(model_path):
+        raise FileNotFoundError(f"PhenotypeSeeker model not found: {model_path}")
+
+    token_id_path = os.path.join(train_dir, 'kmer_token_id.txt')
+    if not os.path.exists(token_id_path):
+        raise FileNotFoundError(f"k-mer token id file not found: {token_id_path}")
+
     o2=open('ps_inf1.txt','w+')
     o3=open('ps_inf2.txt','w+')
     for a in arr_val:
         o2.write(a+'\t'+dval[a]+'\n')
     o2.close()
-    o3.write(drug+'\t'+work_dir+'/PS_out/log_reg_model_'+drug+'.pkl\n')
+    o3.write(drug+'\t'+model_path+'\n')
     o3.close()
     os.system('mv ps_inf1.txt ps_inf2.txt '+work_dir+'/PS_out')
     print(script_dir+'/PhenotypeSeeker/.PSenv/bin/phenotypeseeker prediction '+work_dir+'/PS_out/ps_inf1.txt '+work_dir+'/PS_out/ps_inf2.txt',flush=True)
@@ -276,7 +284,7 @@ def run_ps(intest,label,label2,drug,work_dir):
     if os.path.exists(work_dir+'/PS_out/K-mer_lists'):
         os.system('rm -rf '+work_dir+'/PS_out/K-mer_lists')
     os.system('mv K-mer_lists '+work_dir+'/PS_out')
-    generate_tpsp(work_dir+'/kmer_token_id.txt',work_dir+'/PS_out/K-mer_lists',label2,work_dir+'/strains_test_kmer_token.txt')
+    generate_tpsp(token_id_path,work_dir+'/PS_out/K-mer_lists',label2,work_dir+'/strains_test_kmer_token.txt')
     #os.system('rm -rf K-mer_lists')
     #os.system('rm ps_inf1.txt ps_inf2.txt ')
     os.system('mv predictions_'+drug+'.txt '+work_dir+'/PS_out')
@@ -312,23 +320,59 @@ def cal_len(infile1,infile2):
             ms=token_count
     return ms
 
-def scan_length(odir):
-    o=open(odir+'/longest_len_test.txt','w+')
+def scan_length(test_dir,train_dir):
+    o=open(test_dir+'/longest_len_test.txt','w+')
     o.write('Graph\tPC\tKmer\n')
     #for filename in os.listdir(odir):
     #if not re.search('Fold',filename):continue
-    ls1=cal_len(odir+'/strains_train_sentence.txt',odir+'/strains_test_sentence.txt')
-    ls2=cal_len(odir+'/strains_train_pc_token.txt',odir+'/strains_test_pc_token.txt')
-    ls3=cal_len(odir+'/strains_train_kmer_token.txt',odir+'/strains_test_kmer_token.txt')
+    ls1=cal_len(train_dir+'/strains_train_sentence.txt',test_dir+'/strains_test_sentence.txt')
+    ls2=cal_len(train_dir+'/strains_train_pc_token.txt',test_dir+'/strains_test_pc_token.txt')
+    ls3=cal_len(train_dir+'/strains_train_kmer_token.txt',test_dir+'/strains_test_kmer_token.txt')
     o.write(str(ls1)+'\t'+str(ls2)+'\t'+str(ls3)+'\n')
     
 
 
 #def run(ingenome,label,odir,drug,mfile,intest,label2):
-def run(intest,label2,odir,drug,pc_c,snv_c,kmer_c,mfile,threads=1,feature_limit=None,sentence_limit=None):
-    label=odir+'/train_label.txt'
-    shap_dir = odir + '/shap'
+def run(intest,label2,odir,drug,pc_c,snv_c,kmer_c,mfile,threads=1,feature_limit=None,sentence_limit=None,train_dir=None):
+    work_dir = os.path.abspath(odir)
+    build_dir(work_dir)
+    train_dir = os.path.abspath(train_dir if train_dir else work_dir)
+    label=os.path.join(train_dir,'train_label.txt')
+    if not os.path.exists(label):
+        raise FileNotFoundError(f"Training label file not found: {label}")
+
+    shap_dir = work_dir + '/shap'
     build_dir(shap_dir)
+    train_shap_dir = os.path.join(train_dir, 'shap')
+    if not os.path.isdir(train_shap_dir):
+        raise FileNotFoundError(f"Training SHAP directory not found: {train_shap_dir}")
+    required_shap_files = [
+        'strains_train_sentence_fs_shap_rmf.txt',
+        'strains_train_pc_token_fs_shap_rmf.txt',
+        'strains_train_kmer_token_shap_rmf.txt',
+        'strains_train_sentence_fs_shap_filter.txt',
+        'strains_train_pc_token_fs_shap_filter.txt',
+        'strains_train_kmer_token_shap_filter.txt',
+    ]
+    for rel_path in required_shap_files:
+        candidate = os.path.join(train_shap_dir, rel_path)
+        if not os.path.exists(candidate):
+            raise FileNotFoundError(f"Required SHAP reference file missing: {candidate}")
+    required_feature_files = [
+        os.path.join(train_dir, 'feature_remain_graph.txt'),
+        os.path.join(train_dir, 'feature_remain_pc.txt'),
+        os.path.join(train_dir, 'strains_train_sentence.txt'),
+        os.path.join(train_dir, 'strains_train_pc_token.txt'),
+        os.path.join(train_dir, 'strains_train_kmer_token.txt'),
+        os.path.join(train_dir, 'node_token_match.txt'),
+        os.path.join(train_dir, 'pc_matches.txt'),
+    ]
+    for feature_path in required_feature_files:
+        if not os.path.exists(feature_path):
+            raise FileNotFoundError(f"Required training artifact missing: {feature_path}")
+    shap_sample_file = os.path.join(shap_dir, 'test_samples.tsv')
+    with open(shap_sample_file, 'w') as shap_fh:
+        shap_fh.write('Sample_ID\n')
     dr={}
     val=[]
     for filename in os.listdir(intest):
@@ -339,13 +383,17 @@ def run(intest,label2,odir,drug,pc_c,snv_c,kmer_c,mfile,threads=1,feature_limit=
         dr[pre]=intest+'/'+filename
         val.append(pre)
 
+    with open(shap_sample_file, 'a') as shap_fh:
+        for sid in sorted(val):
+            shap_fh.write(f"{sid}\n")
+
     provided_label_file = bool(label2)
     if provided_label_file:
         label2 = os.path.abspath(label2)
         if not os.path.exists(label2):
             raise FileNotFoundError(f"Provided label file not found: {label2}")
     else:
-        placeholder_label = os.path.join(odir, 'test_labels_placeholder.txt')
+        placeholder_label = os.path.join(work_dir, 'test_labels_placeholder.txt')
         header = 'ID\tLabel'
         if os.path.exists(label):
             with open(label, 'r') as train_label_fh:
@@ -360,11 +408,11 @@ def run(intest,label2,odir,drug,pc_c,snv_c,kmer_c,mfile,threads=1,feature_limit=
         print(f"No test label file supplied. Created placeholder labels at {label2}.", flush=True)
     # Run prodigal and rgi for all input genomes
     print('Run Prodigal and RGI for all input genomes!',flush=True)
-    gdir,pdir=run_prodigal_rgi(dr,odir,threads)
-    gdir=odir+'/Genes_ts'
-    pdir=odir+'/Proteins_ts'
+    gdir,pdir=run_prodigal_rgi(dr,work_dir,threads)
+    gdir=work_dir+'/Genes_ts'
+    pdir=work_dir+'/Proteins_ts'
     #exit()
-    filter_rgi(odir+'/RGI_raw_ts',drug,mfile,odir+'/RGI_ts')
+    filter_rgi(work_dir+'/RGI_raw_ts',drug,mfile,work_dir+'/RGI_ts')
     #exit()
     f=open(label,'r')
     line=f.readline()
@@ -395,7 +443,6 @@ def run(intest,label2,odir,drug,pc_c,snv_c,kmer_c,mfile,threads=1,feature_limit=
         #train=x[train_idx]
         #val=x[val_idx]
         ########### Preprocess ###########
-        work_dir=odir
         #build_dir(work_dir)
         #tem_gt=work_dir+'/genes_train'
         #tem_pt=work_dir+'/proteins_train'
@@ -411,11 +458,11 @@ def run(intest,label2,odir,drug,pc_c,snv_c,kmer_c,mfile,threads=1,feature_limit=
         build_dir(tem_rv)
         
         
-        copy_genome(odir+'/Genes_ts',tem_gv,'gene')
-        copy_genome(odir+'/RGI_ts',tem_rv,'rgi')
+        copy_genome(work_dir+'/Genes_ts',tem_gv,'gene')
+        copy_genome(work_dir+'/RGI_ts',tem_rv,'rgi')
         
 
-        copy_protein(odir+'/Proteins_ts',tem_pv)
+        copy_protein(work_dir+'/Proteins_ts',tem_pv)
         
         
         ########### Graph-based tokens ###########
@@ -425,41 +472,49 @@ def run(intest,label2,odir,drug,pc_c,snv_c,kmer_c,mfile,threads=1,feature_limit=
             gv=work_dir+'/Genomes_test'
             extract(tem_rv,tem_gv,gv)
 
-            graph=work_dir+'/GFA_train_Minimap2'
+            graph=os.path.join(train_dir,'GFA_train_Minimap2')
+            if not os.path.exists(graph):
+                raise FileNotFoundError(f"Training GFA directory not found: {graph}")
 
             align_res=work_dir+'/Align_test_res'
             align(gv,graph,align_res)
 
 
-            generate_at(label2,align_res,work_dir+'/node_token_match.txt',tem_gv,work_dir+'/strains_test_sentence.txt')
+            generate_at(label2,align_res,os.path.join(train_dir,'node_token_match.txt'),tem_gv,work_dir+'/strains_test_sentence.txt')
         
         ############### PC tokens ############
         #ptrain=merge_all_proteins(tem_pt,work_dir,'train')
         if not pc_c==1:
             pval=merge_all_proteins(tem_pv,work_dir,'test')
+            train_merged_proteins = os.path.join(train_dir, 'merged_proteins_train.fa')
+            if not os.path.exists(train_merged_proteins):
+                raise FileNotFoundError(f"Merged training proteins not found: {train_merged_proteins}")
+            train_cluster_file = os.path.join(train_dir, 'merge_train_cdhit.clstr')
+            if not os.path.exists(train_cluster_file):
+                raise FileNotFoundError(f"Training CD-HIT cluster file not found: {train_cluster_file}")
 
-            cls1,cls2=run_cdhit(pval,work_dir)
-            generate_tokens_from_cdhit(work_dir,label2) 
+            run_cdhit(pval,work_dir,train_merged_proteins)
+            generate_tokens_from_cdhit(train_cluster_file,work_dir,label2)
         
         ############### K-mer tokens ##########
         #print(os.path.abspath(work_dir))
         #exit()
         if not kmer_c==1:
-            run_ps(intest,label,label2,drug,os.path.abspath(work_dir))
+            run_ps(intest,label,label2,drug,os.path.abspath(work_dir),train_dir)
         #exit()
         
         #train=x[train_idx]
         #val=x[val_idx]
         sef_test(
             work_dir+'/strains_test_sentence.txt',
-            work_dir+'/feature_remain_graph.txt',
+            os.path.join(train_dir,'feature_remain_graph.txt'),
             work_dir+'/strains_test_sentence_fs.txt',
             feature_limit,
             sentence_limit
         )
         sef_test(
             work_dir+'/strains_test_pc_token.txt',
-            work_dir+'/feature_remain_pc.txt',
+            os.path.join(train_dir,'feature_remain_pc.txt'),
             work_dir+'/strains_test_pc_token_fs.txt',
             feature_limit,
             sentence_limit
@@ -467,28 +522,28 @@ def run(intest,label2,odir,drug,pc_c,snv_c,kmer_c,mfile,threads=1,feature_limit=
         ### For shap
         sef_test(
             work_dir+'/strains_test_sentence_fs.txt',
-            shap_dir+'/strains_train_sentence_fs_shap_rmf.txt',
+            os.path.join(train_shap_dir,'strains_train_sentence_fs_shap_rmf.txt'),
             shap_dir+'/strains_test_sentence_fs_shap_filter.txt',
             None,
             sentence_limit
         )
         sef_test(
             work_dir+'/strains_test_pc_token_fs.txt',
-            shap_dir+'/strains_train_pc_token_fs_shap_rmf.txt',
+            os.path.join(train_shap_dir,'strains_train_pc_token_fs_shap_rmf.txt'),
             shap_dir+'/strains_test_pc_token_fs_shap_filter.txt',
             None,
             sentence_limit
         )
         sef_test(
             work_dir+'/strains_test_kmer_token.txt',
-            shap_dir+'/strains_train_kmer_token_shap_rmf.txt',
+            os.path.join(train_shap_dir,'strains_train_kmer_token_shap_rmf.txt'),
             shap_dir+'/strains_test_kmer_token_shap_filter.txt'
         )
 
         #c+=1
-    scan_length(odir)
+    scan_length(work_dir,train_dir)
     #scan_length_fs(odir)
-    scan_length_fs_shap(shap_dir)
+    scan_length_fs_shap(shap_dir,train_shap_dir)
 
 
 def main():
@@ -500,7 +555,8 @@ def main():
     parser.add_argument('-p','--pc',dest='close_pc',type=int,help="If set to 1, then will skip pc tokens generation step. (Defaut: 0)" ,default=0)
     parser.add_argument('-s','--snv',dest='close_snv',type=int,help="If set to 1, then will skip snv tokens generation step. (Default: 0)",default=0)
     parser.add_argument('-k','--kmer',dest='close_kmer',type=int,help="If set to 1, then will skip k-mer tokens generation step. (Default:0)",default=0)
-    parser.add_argument('-o','--outdir',dest='outdir',type=str,help="Output directory of results. (Note: The output directory of your input test data must match the out directory of your training data.)")
+    parser.add_argument('-o','--outdir',dest='outdir',type=str,help="Output directory for generated test features and reports.")
+    parser.add_argument('--db',dest='db_dir',type=str,help="Directory containing the pre-built training database (default: use the output directory).")
     parser.add_argument('-t','--threads',dest='threads',type=int,help="Number of parallel processes. (Default:1)",default=1)
     parser.add_argument(
         '--feature-limit',
@@ -527,15 +583,23 @@ def main():
     out=args.outdir
     threads=args.threads
     if not out:
-        print('Please provide the output directoy that matches the out directory of your training data!')
+        print('Please provide an output directory for the generated test data!')
         exit()
+    out=os.path.abspath(out)
+    db_dir=args.db_dir
+    if db_dir:
+        db_dir=os.path.abspath(db_dir)
+        if not os.path.isdir(db_dir):
+            raise FileNotFoundError(f"Training database directory not found: {db_dir}")
+    else:
+        db_dir=out
     feature_limit=args.feature_limit
     if feature_limit is not None and feature_limit <= 0:
         feature_limit=None
     sentence_limit=args.sentence_limit
     if sentence_limit is not None and sentence_limit <= 0:
         sentence_limit=None
-    run(infile,lab_file,out,drug,pc_c,snv_c,kmer_c,mfile,threads,feature_limit,sentence_limit)
+    run(infile,lab_file,out,drug,pc_c,snv_c,kmer_c,mfile,threads,feature_limit,sentence_limit,train_dir=db_dir)
 
 if __name__=="__main__":
     sys.exit(main())
